+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.2.5 (TBA)
+
+* Added `PowAssent.Phoenix.ViewHelpers.authorization_link/2` and  `PowAssent.Phoenix.ViewHelpers.deauthorization_link/2`
+* Deprecated `PowAssent.Phoenix.ViewHelpers.provider_link/3`
+
 ## v0.2.4 (2019-04-25)
 
 * Fixed so OAuth 2.0 access token request params are in the POST body in accordance with RFC 6749